Remove repeat classmethod
