+++ py-polars/src/functions/eager.rs
@@ -1,6 +1,6 @@
 use polars::{functions, time};
 use polars_core::datatypes::{TimeUnit, TimeZone};
-use polars_core::prelude::{DataFrame, IntoSeries};
+use polars_core::prelude::*;
 use pyo3::prelude::*;
 
 use crate::conversion::{get_df, get_series, Wrap};
@@ -115,6 +115,41 @@ pub fn hor_concat_df(dfs: &PyAny) -> PyResult<PyDataFrame> {
 
     let df = functions::hor_concat_df(&dfs).map_err(PyPolarsErr::from)?;
     Ok(df.into())
+}
+
+#[pyfunction]
+pub fn repeat_eager(
+    value: Wrap<AnyValue>,
+    n: usize,
+    dtype: Option<Wrap<DataType>>,
+    name: Option<&str>,
+) -> PyResult<PySeries> {
+    let value = value.0;
+    let dtype = match dtype.map(|wrap| wrap.0) {
+        Some(dtype) => dtype,
+        None => match value.dtype() {
+            // Integer inputs that fit in Int32 are parsed as such
+            DataType::Int64 => {
+                let int_value: i64 = value.extract().unwrap();
+                if int_value >= i32::MIN as i64 && int_value <= i32::MAX as i64 {
+                    DataType::Int32
+                } else {
+                    DataType::Int64
+                }
+            }
+            _ => value.dtype(),
+        },
+    };
+    let name = match name {
+        Some(name) => name,
+        None => "",
+    };
+
+    Ok(Series::new(name, &[value])
+        .cast(&dtype)
+        .map_err(PyPolarsErr::from)?
+        .new_from_index(0, n)
+        .into())
 }
 
 #[pyfunction]
