+++ py-polars/src/lib.rs
@@ -87,6 +87,8 @@ fn polars(py: Python, m: &PyModule) -> PyResult<()> {
         .unwrap();
     m.add_wrapped(wrap_pyfunction!(functions::eager::hor_concat_df))
         .unwrap();
+    m.add_wrapped(wrap_pyfunction!(functions::eager::repeat_eager))
+        .unwrap();
     m.add_wrapped(wrap_pyfunction!(functions::eager::time_range))
         .unwrap();
 
@@ -153,7 +155,7 @@ fn polars(py: Python, m: &PyModule) -> PyResult<()> {
         .unwrap();
     m.add_wrapped(wrap_pyfunction!(functions::lazy::reduce))
         .unwrap();
-    m.add_wrapped(wrap_pyfunction!(functions::lazy::repeat))
+    m.add_wrapped(wrap_pyfunction!(functions::lazy::repeat_lazy))
         .unwrap();
     m.add_wrapped(wrap_pyfunction!(functions::lazy::spearman_rank_corr))
         .unwrap();
